Quest system and improvements
Added a quest system that works interactively with the NPC "mother" by pressing the "E" button. A quest is issued, its status and completion are monitored in the console.

By pressing "Q", a panel with the status of the quest appears (minimal information).

Replacing the inventory icon.

Adding highlighting of raised objects (not interactively, I'll finalize it later)